Itelligent http headers filling method (#3)
The headers filling method, used on http GET requests to pipe payloads
is now more intelligent, because instead of fulfilling each header 1by1
since the end of string, and average is now estabilished between the
allowed number of headers and the size each header can theorically carry.
This equilibrates the risks we take on max_headers and those on max_headers_size
when these vars was not especially adapted for the current target.