session MetaDict's deletion now checks item exists
Assuming that MetaDict()'s standard behavior is to delete an item if it
value is set to an empty string or None, no actions are performed if the
item is not already part of self dict based object. Trying to remove
unexisting item raises a KeyError(), and that's obviously not the goal.