core: fix PyPy compat issue with open() behavior
according to: https://pypy.org/compat.html,
'''The main difference that is not going to be fixed is that PyPy does
not support refcounting semantics.'''
It means that: `open('file.txt', 'w').write(data)` won't fill the file
immediately, but only after a certain period of time, when the GC does a
collection.

Fixed all code containing this pattern, by using the `with` keyword.